# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parchi",
    "version": "0.1.0"
  },
  "paths": {
...
    "/batches/{batch_id}/files": {
      "post": {
        "tags": [
          "batches"
        ],
        "summary": "Upload File",
        "description": "Register one file. An identical file is not saved unless confirm_duplicate is true.",
        "operationId": "upload_file_batches__batch_id__files_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_batches__batch_id__files_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/RegisteredResult"
                    },
                    {
                      "$ref": "#/components/schemas/DuplicateResult"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "result",
                    "mapping": {
                      "registered": "#/components/schemas/RegisteredResult",
                      "duplicate": "#/components/schemas/DuplicateResult"
                    }
                  },
                  "title": "Response Upload File Batches  Batch Id  Files Post"
                }
              }
            }
          },
          "200": {
            "description": "Identical file exists; nothing saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/files": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Get Files",
        "operationId": "get_files_files_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FileStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name or reference",
              "title": "Q"
            },
            "description": "Name or reference"
          },
          {
            "name": "attention",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only files a person should look at (the Review queue)",
              "default": false,
              "title": "Attention"
            },
            "description": "Only files a person should look at (the Review queue)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/files/{file_key}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Get One File",
        "description": "One file, by id or reference number, with its receipts, flags and runs.",
        "operationId": "get_one_file_files__file_key__get",
        "parameters": [
          {
            "name": "file_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Remove File",
        "description": "Delete a file with its runs, receipts and flags (D-019). Not while it is processing.",
        "operationId": "remove_file_files__file_key__delete",
        "parameters": [
          {
            "name": "file_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/files/{file_key}/download": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Download File",
        "description": "The original file, under its original name.",
        "operationId": "download_file_files__file_key__download_get",
        "parameters": [
          {
            "name": "file_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Key"
            }
          },
          {
            "name": "inline",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show in the browser (PDFs and images)",
              "default": false,
              "title": "Inline"
            },
            "description": "Show in the browser (PDFs and images)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/files/{file_key}/error-report": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Error Report",
        "description": "A PDF with the file's problems and what each reader found (D-031).",
        "operationId": "error_report_files__file_key__error_report_get",
        "parameters": [
          {
            "name": "file_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/pdf": {}
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/files/{file_key}/receipts": {
      "put": {
        "tags": [
          "files"
        ],
        "summary": "Save Receipts",
        "description": "Save a person's edits as the accepted result; the file becomes resolved (D-030).\nThe reader learns labels it missed (D-048), and the file can be kept as a test.",
        "operationId": "save_receipts_files__file_key__receipts_put",
        "parameters": [
          {
            "name": "file_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedReview"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/files/{file_key}/reject": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Reject",
        "description": "Not a receipt, or a bad scan (D-030 item 4).",
        "operationId": "reject_files__file_key__reject_post",
        "parameters": [
          {
            "name": "file_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_batches__batch_id__files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "confirm_duplicate": {
            "type": "boolean",
            "title": "Confirm Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_batches__batch_id__files_post"
      },
...
      "DuplicateResult": {
        "properties": {
          "result": {
            "type": "string",
            "const": "duplicate",
            "title": "Result",
            "default": "duplicate"
          },
          "duplicate_of": {
            "$ref": "#/components/schemas/FileRef"
          }
        },
        "type": "object",
        "required": [
          "duplicate_of"
        ],
        "title": "DuplicateResult",
        "description": "Nothing was saved. Upload again with confirm_duplicate=true to keep a copy (D-015)."
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorResponse"
      },
...
      "FileDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ref_no": {
            "type": "string",
            "title": "Ref No"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/FileStatus"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "open_flags": {
            "type": "integer",
            "title": "Open Flags",
            "description": "Unresolved problems a person should look at"
          },
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/ReceiptOut"
            },
            "type": "array",
            "title": "Receipts"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/FlagOut"
            },
            "type": "array",
            "title": "Flags"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunOut"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ref_no",
          "original_name",
          "size_bytes",
          "kind",
          "status",
          "error",
          "uploaded_at",
          "duplicate_of",
          "open_flags",
          "receipts",
          "flags",
          "runs"
        ],
        "title": "FileDetail",
        "description": "Everything the Review page shows for one file."
      },
      "FileKind": {
        "type": "string",
        "enum": [
          "excel",
          "csv",
          "pdf_text",
          "pdf_scan",
          "image"
        ],
        "title": "FileKind"
      },
      "FilePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FileSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "FilePage"
      },
      "FileRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ref_no": {
            "type": "string",
            "title": "Ref No"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ref_no",
          "original_name"
        ],
        "title": "FileRef",
        "description": "Enough to name and link to a file."
      },
      "FileStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "parsed",
          "needs_review",
          "ai_processing",
          "flagged",
          "resolved",
          "rejected",
          "failed"
        ],
        "title": "FileStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualEdit": {
        "properties": {
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/ReceiptIn"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Receipts"
          },
          "keep_as_test": {
            "type": "boolean",
            "title": "Keep As Test",
            "description": "Copy the file and this answer to the local test set (D-048)",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "receipts"
        ],
        "title": "ManualEdit"
      },
...
      "RegisteredResult": {
        "properties": {
          "result": {
            "type": "string",
            "const": "registered",
            "title": "Result",
            "default": "registered"
          },
          "file": {
            "$ref": "#/components/schemas/FileSummary"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "RegisteredResult"
      },
      "RejectIn": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Reason",
            "description": "e.g. Not a receipt"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "RejectIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}